k8sconv: a new package for code the converts Kuberentes apis to internal/storage apis
