+++ internal/store/k8sconv/doc.go
@@ -1,0 +1,4 @@
+// Converts Kubernetes data models into our own internal data models.
+//
+// (package name loosely inspired by strconv)
+package k8sconv
